\name{effect.labels}
\alias{effect.labels}
\title{
Create new effect labels for \code{lm}
}
\description{
Alternative notation of effect labels including levels.
}
\usage{
effect.labels(t, data)
}
\arguments{
  \item{t}{
	Terms object.
}
  \item{data}{
	Corresponding model.matrix.
}
}
\value{
  \item{names }{Character vector of effect labels.}
}
\author{
Kristian Hovde Liland
}

\keyword{ effect }
\keyword{ label }
