\name{extend.colnames}
\alias{extend.colnames}
\title{
Extended summary from multinom
}
\description{
Extended summary from multinom
}
\usage{
extend.colnames(object, the.name)
}
\arguments{
  \item{object}{
	object fitted with multinom.
}
  \item{the.name}{
	name to be added to column names.
}
}
\value{
Returns object with new colnames.
}
\author{
Kristian Hovde Liland
}
