\name{summaryMultinom}
\alias{summaryMultinom}
\title{
Extended summary from multinom
}
\description{
Extends the summary function of multinom object from nnet.
}
\usage{
summaryMultinom(object)
}
\arguments{
  \item{object}{
	Object fitted by multinom.
}
}
\value{
Only printing is performed, nothing is returned.
}
\author{
Kristian Hovde Liland
}

\keyword{ Multinomial }
