\name{summaryOrdinal}
\alias{summaryOrdinal}
\title{
Extended summary from ordinal regression
}
\description{
Extended summary from ordinal regression.
}
\usage{
summaryOrdinal(object, digits = max(3, .Options$digits - 3), correlation = FALSE, ...)
}
\arguments{
  \item{object}{
	fitted ordinal regression.
}
  \item{digits}{
	number of digits in output.
}
  \item{correlation}{
	use correlation.
}
  \item{\dots}{
	additional arguments to pass on.
}
}
\value{
Only printing is performed, nothing is returned.
}
\author{
Kristian Hovde Liland
}

\keyword{ Ordinal }
