\name{tally}
\alias{tally}
\title{
Tally of discrete numbers
}
\description{
Tally of discrete numbers
}
\usage{
tally(x)
}
\arguments{
  \item{x}{
	Discrete number vector.
}
}
\value{
Returns the tally.
}
\author{
Kristian Hovde Liland
}

\examples{
tally(c(1,5,1,3,2,5,6,2,2,1,4,3,6))
}

\keyword{ Numbers }
